squashed all changes
bump version up

modify parameters for tests

modify parameters for tests

check slow test

added elbow method

fix test 2

skip slow tests
